Restores keyword params on results page from advanced search
We need to be careful about which parts of blacklight_advanced_search
we override. While working on TD-1219, we also discovered that
constraints would be doubled after using "edit search" and this attempts
to address that by omitting those parameters when JS (and thus the
`chosen` library) is enabled, since these fields are also added by that
method.